new features

+++ pages/edit-deployment.php
@@ -2,10 +2,11 @@
 
 $editId = queryParam('deployment_id');
 $itemId = queryParam('item_id');
-$formMessage = false;
-$deleted = false;
+$formMessage = takeFlash();
 
-if (isset($_POST['edit_deployment_submit'])) {
+$deployment = dbRow('SELECT * FROM inv_deployments WHERE dep_id = :edit_id', ['edit_id' => $editId]);
+
+if ($deployment && isset($_POST['edit_deployment_submit'])) {
     $error = validateDeployment($_POST);
 
     if ($error) {
@@ -21,32 +22,34 @@ if (isset($_POST['edit_deployment_submit'])) {
             ]
         );
 
-        $formMessage = successMessage('Deployment updated!');
+        redirectWith(
+            'index.php?page=edit-deployment&deployment_id=' . urlencode((string)$editId)
+                . '&item_id=' . urlencode((string)$deployment['dep_item_id']),
+            successMessage('Deployment updated!')
+        );
     }
-} elseif (isset($_POST['delete_deployment_submit'])) {
+} elseif ($deployment && isset($_POST['delete_deployment_submit'])) {
     dbRun('DELETE FROM inv_deployments WHERE dep_id = :edit_id LIMIT 1', ['edit_id' => $editId]);
 
-    $formMessage = successMessage('Deployment deleted!');
-    $deleted = true;
+    redirectWith(
+        'index.php?page=view-item&item_id=' . urlencode((string)$deployment['dep_item_id']),
+        successMessage('Deployment deleted!')
+    );
 }
-
-$deployment = dbRow('SELECT * FROM inv_deployments WHERE dep_id = :edit_id', ['edit_id' => $editId]);
 
 pageHeader('Edit Deployment', $itemId ? ['View Item' => 'index.php?page=view-item&item_id=' . $itemId] : []);
 
-if ($deployment) {
-    echo '<form method="post">' . "\n";
+if (!$deployment) {
     formMessage($formMessage);
-    formRow('dep_description', 'Deployment Description',
-        '<input type="text" name="dep_description" value="' . escapeHtml($deployment['dep_description']) . '" />');
-    formRow('dep_quantity', 'Deployment Quantity',
-        '<input type="text" name="dep_quantity" value="' . escapeHtml($deployment['dep_quantity']) . '" />');
-    echo '    <p><input type="submit" name="edit_deployment_submit" value="Save"></p>' . "\n";
-    echo '</form>' . "\n";
-
-    deleteSection('Deployment', 'delete_deployment_submit');
-} elseif ($deleted) {
-    formMessage($formMessage);
-} else {
     echo '<p>No deployment found</p>';
+    return;
 }
+
+echo '<form method="post">' . "\n";
+formMessage($formMessage);
+textField('dep_description', 'Deployment Description', $deployment['dep_description']);
+textField('dep_quantity', 'Deployment Quantity', $deployment['dep_quantity'], 'number', ' min="0"');
+submitButton('edit_deployment_submit');
+echo '</form>' . "\n";
+
+deleteSection('Deployment', 'delete_deployment_submit');
